+++ DiploGM/db/schema.sql
@@ -52,13 +52,14 @@ CREATE TABLE IF NOT EXISTS units (
     FOREIGN KEY (board_id, phase, location) REFERENCES provinces (board_id, phase, province_name),
     FOREIGN KEY (board_id, owner) REFERENCES players (board_id, player_name),
     FOREIGN KEY (board_id, phase, location) REFERENCES retreat_options (board_id, phase, origin));
+
 CREATE TABLE IF NOT EXISTS builds(
     board_id int,
     phase text,
     player text,
     location text,
     order_type text,
-    is_army boolean,
+    unit_type text,
     PRIMARY KEY (board_id, phase, player, location),
     FOREIGN KEY (board_id, phase) REFERENCES boards (board_id, phase),
     FOREIGN KEY (board_id, player) REFERENCES players (board_id, player_name),